fix(graphql-dynamodb-transformer): refactoring for future transforms
Refactoring DynamoDB transformer to move boilerplate into AppSync transformer for future non-model transforms
